add: 入力年齢時点での保険金額を算出するロジックを作成
- 保険金額を算出するためのヘルパーモジュールを作成しました。
- ヘルパーメソッドをコントローラから呼び出して保険金額を計算するようにしました。
