--- app/helpers/birthdays_helper.rb
@@ -1,2 +1,0 @@
-module BirthdaysHelper
-end
