--- app/helpers/insurance_graphs_helper.rb
@@ -1,2 +1,0 @@
-module InsuranceGraphsHelper
-end
