--- app/helpers/insurance_policies_helper.rb
@@ -1,2 +1,0 @@
-module InsurancePoliciesHelper
-end
